% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotargets-package.R
\docType{package}
\name{autotargets-package}
\alias{autotargets}
\alias{autotargets-package}
\title{autotargets: Automatic targets Pipeline Generation}
\description{
Automatically generates targets pipelines by scanning R files for functions and constants, mapping them to tar_target() definitions based on function arguments. No AST analysis or manual DAG specification required.
}
\author{
\strong{Maintainer}: Beau \email{beau@example.com}

}
\keyword{internal}
