% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_assignment_name}
\alias{get_assignment_name}
\title{Get the name from an assignment expression}
\usage{
get_assignment_name(expr)
}
\arguments{
\item{expr}{An assignment expression}
}
\value{
Character string with the assigned name
}
\description{
Get the name from an assignment expression
}
\keyword{internal}
