% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_function_formals}
\alias{get_function_formals}
\title{Get formals from a function definition expression}
\usage{
get_function_formals(expr)
}
\arguments{
\item{expr}{A function definition expression}
}
\value{
List of formal arguments (or NULL if not a function)
}
\description{
Get formals from a function definition expression
}
\keyword{internal}
