% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_constant_def}
\alias{is_constant_def}
\title{Check if expression is a constant (non-function assignment)}
\usage{
is_constant_def(expr)
}
\arguments{
\item{expr}{An R expression}
}
\value{
Logical indicating if expression is a constant assignment
}
\description{
Check if expression is a constant (non-function assignment)
}
\keyword{internal}
