% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_function_def}
\alias{is_function_def}
\title{Check if expression is a function definition}
\usage{
is_function_def(expr)
}
\arguments{
\item{expr}{An R expression}
}
\value{
Logical indicating if expression is a function definition
}
\description{
Check if expression is a function definition
}
\keyword{internal}
