% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_constant_target}
\alias{make_constant_target}
\title{Generate tar_target call string for a constant}
\usage{
make_constant_target(name)
}
\arguments{
\item{name}{Constant name}
}
\value{
Character string with tar_target() call
}
\description{
Generate tar_target call string for a constant
}
\keyword{internal}
