% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_function_target}
\alias{make_function_target}
\title{Generate tar_target call string for a function}
\usage{
make_function_target(name, formals)
}
\arguments{
\item{name}{Function name}

\item{formals}{List of formal arguments}
}
\value{
Character string with tar_target() call
}
\description{
Generate tar_target call string for a function
}
\keyword{internal}
