% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_targets_file}
\alias{write_targets_file}
\title{Write generated targets file with header}
\usage{
write_targets_file(target_lines, path, source_dir = "R/auto/")
}
\arguments{
\item{target_lines}{Character vector of tar_target() call strings}

\item{path}{Output file path}

\item{source_dir}{Source directory that was scanned}
}
\description{
Write generated targets file with header
}
\keyword{internal}
